// src_old/interfaces.ts
export interface Address {
  id?: number;
  street: string;
  city: string;
  state: string;
  country: string;
  postalCode: string;
}

export interface EndUse {
  id?: number;
  name: string;
  description?: string;
}

export interface EndUseWithColor extends EndUse {
  color: string;
}

export interface EndUsesListProps {
  endUses: EndUseWithColor[];
}

export interface Contact {
  id: number;
  name: string;
  designation: string;
  contactNumber: string;
  address?: Address;
  billingAddress: string;
  isMainContact?: boolean;
}

export interface Customer {
  id: number;
  businessName: string;
  contacts: Contact[];
  mainContact?: Contact;
  [key: string]: any;
}
// ...
export interface TableColumn<T> {
  field: keyof T;
  label: string;
  format?: (item: T[keyof T][keyof T[keyof T]] | T[keyof T]) => React.ReactNode;
}

export interface PaginatedResponse<T> {
  data: T[];
  total: number;
  page: number;
  limit: number;
}

export interface GenericTableProps<T> {
  tableName: string;
  columns: TableColumn<T>[];
  fetchData: (
    page: number,
    pageSize: number,
    searchQuery: string
  ) => Promise<PaginatedResponse<T>>;
}

export interface ActionButton {
  icon: React.ReactNode;
  onClick: (id: number, item?: any) => void;
  title?: string;
  disabled?: boolean;
}

export interface SearchBarProps {
  onSearch: (value: string) => void;
}

export interface AdditionalColumn<T> {
  type: 'radio' | 'checkbox';
  onChange: (itemId: number, checked: boolean) => void;
  valueGetter: (item: T) => boolean;
  columnName: string;
}

export interface ModalFormProps {
  labels: string[];
  buttonText: string;
  onConfirmClick: () => void;
  onCancelClick: () => void;
  onModalDataSubmit: (data: string[]) => void;
}

export interface FieldConfig<T> {
  label: string;
  name: keyof T | string;
  type?: string;
}